% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchadd.R
\name{bench.addmethod}
\alias{bench.addmethod}
\title{add method to BenchmarkingR project}
\usage{
bench.addmethod(bench.proj, method.func, method.name, method.description = "",
  replace = FALSE)
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

