% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchadd.R
\name{bench.addparameter}
\alias{bench.addparameter}
\title{add parameter setting for data.name and method.name}
\usage{
bench.addparameter(bench.proj, data.name, method.name, parameter)
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

