% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchsampler.R
\name{bench.addsampler}
\alias{bench.addsampler}
\title{add sampler to BenchmarkingR project}
\usage{
bench.addsampler(bench.proj, sampler.func, sampler.name,
  default.parameter = list(), sampler.description = "")
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

