% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchget.R
\name{bench.getdata.exploration}
\alias{bench.getdata.exploration}
\title{Getter a data.exploration}
\usage{
bench.getdata.exploration(bench.proj, data.name)
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

