% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchplot.R
\name{bench.heatmap}
\alias{bench.heatmap}
\title{Compute a heatmap of cerrelation between methods pvalue}
\usage{
bench.heatmap(bench.proj, data.name, methods.name)
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

