% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchplot.R
\name{bench.plotfdr_power}
\alias{bench.plotfdr_power}
\title{Plot fdr and power of method for this data.name. Use B-H procedure for controlling the FDR}
\usage{
bench.plotfdr_power(bench.proj, data.name, methods.name, fdr_lim = c(0, 0.25))
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

