% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bench.R
\name{bench}
\alias{bench}
\title{Create a new benchmarkingR project dirrectory if the BenchmarkingR directory does not exist.
If the directory exits retrieve the BenchmarkingR project}
\usage{
bench(dir.name = NULL, new = FALSE, patatorDir = NULL)
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

