% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bench.R
\name{bench.remove}
\alias{bench.remove}
\title{remove specified already computed benchmark}
\usage{
bench.remove(bench.proj, data.name, method.name)
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

