% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchrun.R
\name{bench.run}
\alias{bench.run}
\title{run bencmarking}
\usage{
bench.run(bench.proj, data.name = NULL, method.name = NULL, again = FALSE,
  parameter = list())
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

