% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchsampler.R
\name{bench.runsampler}
\alias{bench.runsampler}
\title{add sampler to BenchmarkingR project}
\usage{
bench.runsampler(bench.proj, sampler.name, repetition, parameter)
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

