% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchadd.R
\name{bench.adddataset}
\alias{bench.adddataset}
\title{add a data set to BenchmarkingR project}
\usage{
bench.adddataset(bench.proj, data.G, data.X, data.outlier, data.name,
  data.description = "", replace = FALSE)
}
\value{
TODO
}
\description{
TODO
}
\examples{
TODO
}

